+++ packages/contact-center/cc-widgets/src/wc.ts
@@ -2,8 +2,22 @@ import r2wc from '@r2wc/react-to-web-component';
 import {StationLogin} from '@webex/cc-station-login';
 import {UserState} from '@webex/cc-user-state';
 import store from '@webex/cc-store';
+import {TaskList, IncomingTask} from '@webex/cc-task';
 
 const WebUserState = r2wc(UserState);
+const WebIncomingTask = r2wc(IncomingTask, {
+  props: {
+    onAccepted: 'function',
+    onDeclined: 'function',
+  },
+});
+
+const WebTaskList = r2wc(TaskList, {
+  props: {
+    onTaskAccepted: 'function',
+    onTaskDeclined: 'function',
+  },
+});
 
 const WebStationLogin = r2wc(StationLogin, {
   props: {
@@ -18,6 +32,14 @@ if (!customElements.get('widget-cc-user-state')) {
 
 if (!customElements.get('widget-cc-station-login')) {
   customElements.define('widget-cc-station-login', WebStationLogin);
+}
+
+if (!customElements.get('widget-cc-incoming-task')) {
+  customElements.define('widget-cc-incoming-task', WebIncomingTask);
+}
+
+if (!customElements.get('widget-cc-task-list')) {
+  customElements.define('widget-cc-task-list', WebTaskList);
 }
 
 export {store};