+++ packages/contact-center/CHANGELOG.md
@@ -1,3 +1,17 @@
+# [1.28.0-ccwidgets.7](https://github.com/webex/widgets/compare/v1.28.0-ccwidgets.6...v1.28.0-ccwidgets.7) (2025-01-08)
+
+
+### Features
+
+* **widgets:** shifted-timer-to-worker ([#352](https://github.com/webex/widgets/issues/352)) ([c06fe9c](https://github.com/webex/widgets/commit/c06fe9c208914d86281950e9a8373ce778d8f0f2))
+
+# [1.28.0-ccwidgets.6](https://github.com/webex/widgets/compare/v1.28.0-ccwidgets.5...v1.28.0-ccwidgets.6) (2025-01-08)
+
+
+### Features
+
+* taskList and IncomingTask widgets added ([#348](https://github.com/webex/widgets/issues/348)) ([ce3a619](https://github.com/webex/widgets/commit/ce3a619dbbb1e2e5321a94a7ce2798f6217326b9))
+
 # [1.28.0-ccwidgets.5](https://github.com/webex/widgets/compare/v1.28.0-ccwidgets.4...v1.28.0-ccwidgets.5) (2024-12-23)